# src/main/resources/openapi.yaml
openapi: 3.1.0
info:
  title: api-lifecare-integrator
  contact: {}
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
  version: "1.0"
servers:
- url: http://localhost:61256
  description: Generated server url
...
paths:
...
  /{municipalityId}/services:
    get:
      tags:
      - Case data
      description: Get the services registered on a person in the given period
      operationId: getServices
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: partyId
        in: query
        description: Party id of the person
        required: true
        schema:
          type: string
        example: 81471222-5798-11e9-ae24-57fa13b361e1
      - name: from
        in: query
        description: Start of the period
        required: false
        schema:
          type: string
          format: date
        example: 2025-01-01
      - name: to
        in: query
        description: End of the period
        required: false
        schema:
          type: string
          format: date
        example: 2026-12-31
      - name: page
        in: query
        description: Page number
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
        example: 1
      - name: pageSize
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
          maximum: 1000
        example: 20
      - name: ascending
        in: query
        description: Sort order
        required: false
        schema:
          type: boolean
        example: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseCaseService"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/resource-allocations:
    get:
      tags:
      - Case data
      description: Get the resource allocations registered on a person in the given
        period
      operationId: getResourceAllocations
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: partyId
        in: query
        description: Party id of the person
        required: true
        schema:
          type: string
        example: 81471222-5798-11e9-ae24-57fa13b361e1
      - name: from
        in: query
        description: Start of the period
        required: false
        schema:
          type: string
          format: date
        example: 2025-01-01
      - name: to
        in: query
        description: End of the period
        required: false
        schema:
          type: string
          format: date
        example: 2026-12-31
      - name: page
        in: query
        description: Page number
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
        example: 1
      - name: pageSize
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
          maximum: 1000
        example: 20
      - name: ascending
        in: query
        description: Sort order
        required: false
        schema:
          type: boolean
        example: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseResourceAllocation"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
  /{municipalityId}/payments:
    get:
      tags:
      - Case data
      description: Get the payments registered on a person in the given period
      operationId: getPayments
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: partyId
        in: query
        description: Party id of the person
        required: true
        schema:
          type: string
        example: 81471222-5798-11e9-ae24-57fa13b361e1
      - name: from
        in: query
        description: Start of the period
        required: false
        schema:
          type: string
          format: date
        example: 2025-01-01
      - name: to
        in: query
        description: End of the period
        required: false
        schema:
          type: string
          format: date
        example: 2026-12-31
      - name: page
        in: query
        description: Page number
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
        example: 1
      - name: pageSize
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
          maximum: 1000
        example: 20
      - name: ascending
        in: query
        description: Sort order
        required: false
        schema:
          type: boolean
        example: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponsePayment"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/investigations:
    get:
      tags:
      - Case data
      description: Get the investigations registered on a person in the given period
      operationId: getInvestigations
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: partyId
        in: query
        description: Party id of the person
        required: true
        schema:
          type: string
        example: 81471222-5798-11e9-ae24-57fa13b361e1
      - name: from
        in: query
        description: Start of the period
        required: false
        schema:
          type: string
          format: date
        example: 2025-01-01
      - name: to
        in: query
        description: End of the period
        required: false
        schema:
          type: string
          format: date
        example: 2026-12-31
      - name: page
        in: query
        description: Page number
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
        example: 1
      - name: pageSize
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
          maximum: 1000
        example: 20
      - name: ascending
        in: query
        description: Sort order
        required: false
        schema:
          type: boolean
        example: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseInvestigation"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
  /{municipalityId}/executions:
    get:
      tags:
      - Case data
      description: Get the executions registered on a person in the given period
      operationId: getExecutions
      parameters:
      - name: municipalityId
        in: path
        description: Municipality id
        required: true
        schema:
          type: string
        example: 2281
      - name: partyId
        in: query
        description: Party id of the person
        required: true
        schema:
          type: string
        example: 81471222-5798-11e9-ae24-57fa13b361e1
      - name: from
        in: query
        description: Start of the period
        required: false
        schema:
          type: string
          format: date
        example: 2025-01-01
      - name: to
        in: query
        description: End of the period
        required: false
        schema:
          type: string
          format: date
        example: 2026-12-31
      - name: page
        in: query
        description: Page number
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
        example: 1
      - name: pageSize
        in: query
        description: Page size
        required: false
        schema:
          type: integer
          format: int32
          exclusiveMinimum: 0
          maximum: 1000
        example: 20
      - name: ascending
        in: query
        description: Sort order
        required: false
        schema:
          type: boolean
        example: true
      responses:
        "200":
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseExecution"
        "404":
          description: Not found
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
        "400":
          description: Bad request
          content:
            application/problem+json:
              schema:
                oneOf:
                - $ref: "#/components/schemas/Problem"
                - $ref: "#/components/schemas/ConstraintViolationProblem"
        "500":
          description: Internal server error
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
...
components:
  schemas:
    Problem:
      type: object
      properties:
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
...
    ConstraintViolationProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        status:
          type: integer
          format: int32
        violations:
          type: array
          items:
            $ref: "#/components/schemas/Violation"
        title:
          type: string
        causeAsProblem:
          $ref: "#/components/schemas/ThrowableProblem"
        detail:
          type: string
        instance:
          type: string
          format: uri
    ThrowableProblem:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
          format: int32
        detail:
          type: string
        instance:
          type: string
          format: uri
        causeAsProblem: {}
    Violation:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    CaseService:
      type: object
      description: A service (insats) from the Lifecare family care system
      properties:
        id:
          type: integer
          format: int32
          description: The service id in the source system
          examples:
          - 12345
        type:
          type: string
          description: The service type
          examples:
          - Ekonomiskt bistånd
        organization:
          type: string
          description: The organization handling the service
          examples:
          - IFO Vuxen
        startDate:
          type: string
          format: date
          description: The date the service starts
          examples:
          - 2026-05-01
        endDate:
          type: string
          format: date
          description: The date the service ends
          examples:
          - 2026-10-31
        caseworker:
          type: string
          description: The name of the caseworker
          examples:
          - Anna Andersson
        coCaseworker:
          type: string
          description: The name of the co-caseworker
          examples:
          - Bengt Bengtsson
        investigationId:
          type: integer
          format: int32
          description: The id of the investigation the service belongs to
          examples:
          - 23456
        decisionId:
          type: integer
          format: int32
          description: The id of the decision the service belongs to
          examples:
          - 34567
        applicant:
          type: string
          description: The name of the applicant
          examples:
          - Cecilia Cedersson
        coApplicant:
          type: string
          description: The name of the co-applicant
          examples:
          - David Davidsson
        persons:
          type: array
          description: Names of the persons the service concerns
          examples:
          - - Anna Andersson
          items:
            type: string
    PagedResponseCaseService:
      type: object
      description: "Paged response wrapper, mirroring the pagination of the underlying\
        \ Lifecare family care API"
      properties:
        page:
          type: integer
          format: int32
          description: The current page (one-based)
          examples:
          - 1
        pageSize:
          type: integer
          format: int32
          description: The number of records per page
          examples:
          - 20
        totalPages:
          type: integer
          format: int32
          description: The total number of pages
          examples:
          - 5
        totalRecords:
          type: integer
          format: int32
          description: The total number of records
          examples:
          - 98
        results:
          type: array
          description: The records on the current page
          items:
            $ref: "#/components/schemas/CaseService"
    PagedResponseResourceAllocation:
      type: object
      description: "Paged response wrapper, mirroring the pagination of the underlying\
        \ Lifecare family care API"
      properties:
        page:
          type: integer
          format: int32
          description: The current page (one-based)
          examples:
          - 1
        pageSize:
          type: integer
          format: int32
          description: The number of records per page
          examples:
          - 20
        totalPages:
          type: integer
          format: int32
          description: The total number of pages
          examples:
          - 5
        totalRecords:
          type: integer
          format: int32
          description: The total number of records
          examples:
          - 98
        results:
          type: array
          description: The records on the current page
          items:
            $ref: "#/components/schemas/ResourceAllocation"
    ResourceAllocation:
      type: object
      description: A resource allocation from the Lifecare family care system
      properties:
        id:
          type: integer
          format: int32
          description: The resource allocation id in the source system
          examples:
          - 12345
        startDate:
          type: string
          format: date
          description: The date the resource allocation starts
          examples:
          - 2026-05-01
        endDate:
          type: string
          format: date
          description: The date the resource allocation ends
          examples:
          - 2026-10-31
        percent:
          type: integer
          format: int32
          description: The allocation percentage
          examples:
          - 100
        resource:
          type: string
          description: The allocated resource
          examples:
          - Anna Andersson
        resourceType:
          type: string
          description: The type of the allocated resource
          examples:
          - Handläggare
        serviceId:
          type: integer
          format: int32
          description: The id of the service the resource allocation belongs to
          examples:
          - 23456
...
    PagedResponsePayment:
      type: object
      description: "Paged response wrapper, mirroring the pagination of the underlying\
        \ Lifecare family care API"
      properties:
        page:
          type: integer
          format: int32
          description: The current page (one-based)
          examples:
          - 1
        pageSize:
          type: integer
          format: int32
          description: The number of records per page
          examples:
          - 20
        totalPages:
          type: integer
          format: int32
          description: The total number of pages
          examples:
          - 5
        totalRecords:
          type: integer
          format: int32
          description: The total number of records
          examples:
          - 98
        results:
          type: array
          description: The records on the current page
          items:
            $ref: "#/components/schemas/Payment"
...
    Investigation:
      type: object
      description: An investigation from the Lifecare family care system
      properties:
        id:
          type: integer
          format: int32
          description: The investigation id in the source system
          examples:
          - 12345
        type:
          type: string
          description: The investigation type
          examples:
          - Ekonomiskt bistånd
        fromDate:
          type: string
          format: date
          description: The date the investigation starts
          examples:
          - 2026-05-01
        toDate:
          type: string
          format: date
          description: The date the investigation ends
          examples:
          - 2026-10-31
        caseworker:
          type: string
          description: The name of the caseworker
          examples:
          - Anna Andersson
        organization:
          type: string
          description: The organization handling the investigation
          examples:
          - IFO Vuxen
        dossierType:
          type: string
          description: The dossier type
          examples:
          - Vuxen
        applicant:
          type: string
          description: The name of the applicant
          examples:
          - Bengt Bengtsson
        coApplicant:
          type: string
          description: The name of the co-applicant
          examples:
          - Cecilia Cedersson
        persons:
          type: array
          description: The persons included in the investigation
          items:
            $ref: "#/components/schemas/InvestigationPerson"
    InvestigationPerson:
      type: object
      description: A person included in an investigation
      properties:
        name:
          type: string
          description: The name of the person
          examples:
          - Anna Andersson
        coApplicant:
          type: boolean
          description: Whether the person is a co-applicant
          examples:
          - "false"
    PagedResponseInvestigation:
      type: object
      description: "Paged response wrapper, mirroring the pagination of the underlying\
        \ Lifecare family care API"
      properties:
        page:
          type: integer
          format: int32
          description: The current page (one-based)
          examples:
          - 1
        pageSize:
          type: integer
          format: int32
          description: The number of records per page
          examples:
          - 20
        totalPages:
          type: integer
          format: int32
          description: The total number of pages
          examples:
          - 5
        totalRecords:
          type: integer
          format: int32
          description: The total number of records
          examples:
          - 98
        results:
          type: array
          description: The records on the current page
          items:
            $ref: "#/components/schemas/Investigation"
...
    PagedResponseExecution:
      type: object
      description: "Paged response wrapper, mirroring the pagination of the underlying\
        \ Lifecare family care API"
      properties:
        page:
          type: integer
          format: int32
          description: The current page (one-based)
          examples:
          - 1
        pageSize:
          type: integer
          format: int32
          description: The number of records per page
          examples:
          - 20
        totalPages:
          type: integer
          format: int32
          description: The total number of pages
          examples:
          - 5
        totalRecords:
          type: integer
          format: int32
          description: The total number of records
          examples:
          - 98
        results:
          type: array
          description: The records on the current page
          items:
            $ref: "#/components/schemas/Execution"
...
  securitySchemes: {}